+++ setup.py
@@ -8,7 +8,7 @@ long_description = open('README.md', 'r', encoding="UTF-8").read()
 
 setup(
   name = 'SCALECAST',
-  version = '0.16.1',
+  version = '0.16.3',
   license='MIT',
   description="The practitioner's time series forecasting library",
   long_description=long_description,
