+++ src/scalecast/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.16.1"
+__version__ = "0.16.3"
